Initial Commit
This commit introduces key files for the setup and configuration of a new Next.js application. This includes the main app layout file (layout.tsx), Next.js and TypeScript configuration files, and initial package.json for dependencies and scripts configuration. Also included are README and .env.example files for basic project information and environment setup respectively. The addition of these files forms the foundation for the project's development.