Create server-side application and delete default Next.js images
This commit introduces the structure for the server-side of the application, adding new files and setting up their initial configurations.
A key change is the creation and configuration of the `Cargo.toml` file, which lists the necessary dependencies for the Rust Web server.
Additionally, files related to dealing with authentication requests (login, logout, auth, register) are created under 'server/src/routes/api/auth.
The main server file (`main.rs`) is also implemented to establish the connection to the database and handle some potential HTTP responses.
Lastly, the default `next.svg` and `vercel.svg` images are removed from the public directory.