Upgrade tracing dependencies and add HTTP tracing layer
Updated the 'tracing-subscriber' dependency to version '0.3.18' and 'tracing' package to version '0.1.37'. Also added the HTTP tracing layer to the router and included formatted, thread naming, and level limited tracing in the main function. The profile model was adjusted and no longer includes an ID field, instead the `user_id` is now the primary key. Furthermore, made several changes to the profile update feature for better pronoun management.