Refactor authentication system and update overall layout
The authentication system has been restructured to include separate login forms for email and passkey. This includes the deletion of the original loginForm component, the creation of new ones, and changes to the page layout to accommodate these modifications. Furthermore, some components have been moved into a shared 'auth' directory for reusability.