Update dependencies and improve contact information rendering
Upgraded several dependencies in 'pnpm-lock.yaml' to their latest stable versions. This includes '@tanstack/react-query', '@types/node' and '@types/react' among others. Also, modified 'page.tsx' to ensure that the 'last' prop is correctly assigned to each contact information item. The 'last' prop is used to determine if the current item is the last in the list of contact information.